Extract setting APIs.
