+++ binding.gyp
@@ -72,6 +72,7 @@
         'src/api/steam_api_auth.cc',
         'src/api/steam_api_cloud.cc',
         'src/api/steam_api_friends.cc',
+        'src/api/steam_api_settings.cc',
         'src/api/steam_api_registry.h',
         'src/api/steam_api_workshop.cc',
         'src/greenworks_api.cc',
